+++ PaymentsSharing/Payments/Payment.cs
@@ -6,6 +6,6 @@ internal readonly record struct Payment(
     DateTime CreatedAt,
     IReadOnlyCollection<Person> Payers,
     IReadOnlyCollection<Person> Consumers,
-    decimal Amount,
-    decimal? AmountForMeat = null,
+    uint Amount,
+    uint? AmountForMeat = null,
     string Description = "");
